Move `RegistrationForm` and `LogInForm` to `src/components/forms/auth`
